requirements.txt
google-genai==1.15.0
python-dotenv==1.1.0
python-docx==1.1.2